requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
duckduckgo-search
sentence-transformers
pinecone-client
PyJWT==2.8.0
boto3
redis
